Refactor tasks, change order
